CODE FROM BAG DAY

+++ src/org/usfirst/frc4946/Launcher.java
@@ -59,7 +59,7 @@ public class Launcher {
             double bottomDifference = m_launcherBottomCounter.getRPM() - m_topVelocity.getTargetSpeed();
             
             m_driverStationLCD.println(RobotConstants.LCD_LAUNCHER, 1,
-                    "TCmd:" + m_topVelocity.getTargetSpeed() 
+                    "TCmd:" +(int) m_topVelocity.getTargetSpeed() 
                     + "T:" + (int) topDifference
                     + "|B:" + (int) bottomDifference
                     + "                         ");
